move Body to App
remove Body component
